+++ src/configure/activate.ts
@@ -3,7 +3,7 @@ import * as vscode from 'vscode';
 import { configurePipeline } from './configure';
 import { Messages } from './resources/messages';
 import { AzureAccountExtensionExports, extensionVariables } from './model/models';
-import { telemetryHelper } from './helper/telemetryHelper';
+import { telemetryHelper } from '../helpers/telemetryHelper';
 
 export async function activateConfigurePipeline(): Promise<void> {
     let azureAccountExtension = vscode.extensions.getExtension("ms-vscode.azure-account");
